Add files via upload
This assignment is a programming assignment. This means that your submission will consist of a link to your personal public GitHub repository with a properly accomplished assignment template that will be graded via automated testing. 

Find the appropriate template in the class GitHubLinks to an external site.. Accomplish it appropriately and upload the completed template to your personal public GitHub repository with the same file name. Do not change the function signatures; we will reference them to grade your work.

Use the template `mod-4-assignment-templates/mod-4-intermediate.py` for this specific assignment.